Removed duplication from PatternDSL
Removed other duplication
